feat: route approval copy generation via daemon dependency (M1)
