+++ assistant/src/runtime/routes/channel-routes.ts
@@ -54,6 +54,7 @@ import type { RunOrchestrator } from '../run-orchestrator.js';
 import type {
   MessageProcessor,
   RuntimeAttachmentMetadata,
+  ApprovalCopyGenerator,
 } from '../http-types.js';
 import type { GuardianRuntimeContext } from '../../daemon/session-runtime-assembly.js';
 import { composeApprovalMessageGenerative } from '../approval-message-composer.js';
@@ -158,6 +159,7 @@ interface DeliverGeneratedApprovalPromptParams {
   prompt: ChannelApprovalPrompt;
   uiMetadata: ApprovalUIMetadata;
   messageContext: ApprovalMessageContext;
+  approvalCopyGenerator?: ApprovalCopyGenerator;
 }
 
 /**
@@ -176,6 +178,7 @@ async function deliverGeneratedApprovalPrompt(params: DeliverGeneratedApprovalPr
     prompt,
     uiMetadata,
     messageContext,
+    approvalCopyGenerator,
   } = params;
   const keywords = requiredDecisionKeywords(uiMetadata.actions);
 
@@ -183,6 +186,7 @@ async function deliverGeneratedApprovalPrompt(params: DeliverGeneratedApprovalPr
     const richText = await composeApprovalMessageGenerative(
       { ...messageContext, channel: sourceChannel, richUi: true },
       { fallbackText: prompt.promptText },
+      approvalCopyGenerator,
     );
 
     try {
@@ -205,6 +209,7 @@ async function deliverGeneratedApprovalPrompt(params: DeliverGeneratedApprovalPr
     const plainTextFallback = await composeApprovalMessageGenerative(
       { ...messageContext, channel: sourceChannel, richUi: false },
       { fallbackText: prompt.plainTextFallback, requiredKeywords: keywords },
+      approvalCopyGenerator,
     );
 
     // Embed the run reference so plain-text replies can disambiguate when
@@ -230,6 +235,7 @@ async function deliverGeneratedApprovalPrompt(params: DeliverGeneratedApprovalPr
   const plainText = await composeApprovalMessageGenerative(
     { ...messageContext, channel: sourceChannel, richUi: false },
     { fallbackText: prompt.plainTextFallback, requiredKeywords: keywords },
+    approvalCopyGenerator,
   );
 
   // Embed the run reference for disambiguation in multi-pending scenarios.
@@ -335,6 +341,7 @@ export async function handleChannelInbound(
   runOrchestrator?: RunOrchestrator,
   assistantId: string = 'self',
   gatewayOriginSecret?: string,
+  approvalCopyGenerator?: ApprovalCopyGenerator,
 ): Promise<Response> {
   // Reject requests that lack valid gateway-origin proof. This ensures
   // channel inbound messages can only arrive via the gateway (which
@@ -545,12 +552,12 @@ export async function handleChannelInbound(
         ? await composeApprovalMessageGenerative({
           scenario: 'guardian_verify_success',
           channel: sourceChannel,
-        })
+        }, {}, approvalCopyGenerator)
         : await composeApprovalMessageGenerative({
           scenario: 'guardian_verify_failed',
           channel: sourceChannel,
           failureReason: stripVerificationFailurePrefix(verifyResult.reason),
-        });
+        }, {}, approvalCopyGenerator);
 
       try {
         await deliverChannelReply(replyCallbackUrl, {
@@ -773,6 +780,7 @@ export async function handleChannelInbound(
       orchestrator: runOrchestrator,
       guardianCtx,
       assistantId,
+      approvalCopyGenerator,
     });
 
     if (approvalResult.handled) {
@@ -854,6 +862,7 @@ export async function handleChannelInbound(
         metadataUxBrief,
         commandIntent,
         sourceLanguageCode,
+        approvalCopyGenerator,
       });
     } else {
       // Fire-and-forget: process the message and deliver the reply in the background.
@@ -1006,6 +1015,7 @@ interface ApprovalProcessingParams {
   metadataUxBrief?: string;
   commandIntent?: Record<string, unknown>;
   sourceLanguageCode?: string;
+  approvalCopyGenerator?: ApprovalCopyGenerator;
 }
 
 /**
@@ -1037,6 +1047,7 @@ function processChannelMessageWithApprovals(params: ApprovalProcessingParams): v
     metadataUxBrief,
     commandIntent,
     sourceLanguageCode,
+    approvalCopyGenerator,
   } = params;
 
   const isNonGuardian = guardianCtx.actorRole === 'non-guardian';
@@ -1135,6 +1146,7 @@ function processChannelMessageWithApprovals(params: ApprovalProcessingParams): v
                 toolName: pending[0].toolName,
                 requesterIdentifier: guardianCtx.requesterIdentifier ?? 'Unknown user',
               },
+              approvalCopyGenerator,
             });
 
             if (guardianNotified) {
@@ -1160,7 +1172,7 @@ function processChannelMessageWithApprovals(params: ApprovalProcessingParams): v
                   scenario: 'guardian_request_forwarded',
                   toolName: pending[0].toolName,
                   channel: sourceChannel,
-                });
+                }, {}, approvalCopyGenerator);
                 await deliverChannelReply(replyCallbackUrl, {
                   chatId: guardianCtx.requesterChatId ?? externalChatId,
                   text: forwardedText,
@@ -1188,6 +1200,7 @@ function processChannelMessageWithApprovals(params: ApprovalProcessingParams): v
                   scenario: 'standard_prompt',
                   toolName: pending[0].toolName,
                 },
+                approvalCopyGenerator,
               });
               if (delivered) {
                 hasPostDecisionDelivery = true;
@@ -1326,6 +1339,7 @@ interface ApprovalInterceptionParams {
   orchestrator: RunOrchestrator;
   guardianCtx: GuardianContext;
   assistantId: string;
+  approvalCopyGenerator?: ApprovalCopyGenerator;
 }
 
 interface ApprovalInterceptionResult {
@@ -1358,6 +1372,7 @@ async function handleApprovalInterception(
     orchestrator,
     guardianCtx,
     assistantId,
+    approvalCopyGenerator,
   } = params;
 
   // ── Guardian approval decision path ──
@@ -1397,7 +1412,7 @@ async function handleApprovalInterception(
             scenario: 'guardian_disambiguation',
             pendingCount: allPending.length,
             channel: sourceChannel,
-          });
+          }, {}, approvalCopyGenerator);
           await deliverChannelReply(replyCallbackUrl, {
             chatId: externalChatId,
             text: disambiguationText,
@@ -1434,7 +1449,7 @@ async function handleApprovalInterception(
           const mismatchText = await composeApprovalMessageGenerative({
             scenario: 'guardian_identity_mismatch',
             channel: sourceChannel,
-          });
+          }, {}, approvalCopyGenerator);
           await deliverChannelReply(replyCallbackUrl, {
             chatId: externalChatId,
             text: mismatchText,
@@ -1476,7 +1491,7 @@ async function handleApprovalInterception(
             decision: decision.action === 'reject' ? 'denied' : 'approved',
             toolName: guardianApproval.toolName,
             channel: sourceChannel,
-          });
+          }, {}, approvalCopyGenerator);
           try {
             await deliverChannelReply(replyCallbackUrl, {
               chatId: guardianApproval.requesterChatId,
@@ -1528,6 +1543,7 @@ async function handleApprovalInterception(
               channel: sourceChannel,
               toolName: pendingInfo[0].toolName,
             },
+            approvalCopyGenerator,
           });
           if (!delivered) {
             log.error(
@@ -1590,7 +1606,7 @@ async function handleApprovalInterception(
           const pendingText = await composeApprovalMessageGenerative({
             scenario: 'request_pending_guardian',
             channel: sourceChannel,
-          });
+          }, {}, approvalCopyGenerator);
           await deliverChannelReply(replyCallbackUrl, {
             chatId: externalChatId,
             text: pendingText,
@@ -1622,7 +1638,7 @@ async function handleApprovalInterception(
             scenario: 'guardian_expired_requester',
             toolName: pending[0].toolName,
             channel: sourceChannel,
-          });
+          }, {}, approvalCopyGenerator);
           await deliverChannelReply(replyCallbackUrl, {
             chatId: externalChatId,
             text: expiredText,
@@ -1704,6 +1720,7 @@ async function handleApprovalInterception(
           channel: sourceChannel,
           toolName: pending[0].toolName,
         },
+        approvalCopyGenerator,
       });
       if (!delivered) {
         log.error({ conversationId, externalChatId }, 'Failed to deliver approval reminder');
@@ -1879,6 +1896,7 @@ export function sweepExpiredGuardianApprovals(
   orchestrator: RunOrchestrator,
   gatewayBaseUrl: string,
   bearerToken?: string,
+  approvalCopyGenerator?: ApprovalCopyGenerator,
 ): void {
   const expired = getExpiredPendingApprovals();
   for (const approval of expired) {
@@ -1901,7 +1919,7 @@ export function sweepExpiredGuardianApprovals(
         scenario: 'guardian_expired_requester',
         toolName: approval.toolName,
         channel: approval.channel,
-      });
+      }, {}, approvalCopyGenerator);
       await deliverChannelReply(deliverUrl, {
         chatId: approval.requesterChatId,
         text: requesterText,
@@ -1918,7 +1936,7 @@ export function sweepExpiredGuardianApprovals(
         toolName: approval.toolName,
         requesterIdentifier: approval.requesterExternalUserId,
         channel: approval.channel,
-      });
+      }, {}, approvalCopyGenerator);
       await deliverChannelReply(deliverUrl, {
         chatId: approval.guardianChatId,
         text: guardianText,
@@ -1943,11 +1961,12 @@ export function startGuardianExpirySweep(
   orchestrator: RunOrchestrator,
   gatewayBaseUrl: string,
   bearerToken?: string,
+  approvalCopyGenerator?: ApprovalCopyGenerator,
 ): void {
   if (expirySweepTimer) return;
   expirySweepTimer = setInterval(() => {
     try {
-      sweepExpiredGuardianApprovals(orchestrator, gatewayBaseUrl, bearerToken);
+      sweepExpiredGuardianApprovals(orchestrator, gatewayBaseUrl, bearerToken, approvalCopyGenerator);
     } catch (err) {
       log.error({ err }, 'Guardian expiry sweep failed');
     }
